+++ src/lib/actions/index.ts
@@ -1,7 +1,7 @@
 export { default as readUser } from './readUser';
 
 export { default as readPosts } from './readPosts';
-export { default as readPostWithUserAndCommentsCountAndReactionsCounts } from './readPostWithUserAndCommentsCountAndReactionsCounts';
+export { default as readPostWithUserAndCommentsCountAndReactionsCountsAndUserReaction } from './readPostWithUserAndCommentsCountAndReactionsCountsAndUserReaction';
 
 export { default as readComments } from './readComments';
 export { default as readCommentWithUserAndRepliesCount } from './readCommentWithUserAndRepliesCount';
